chore: rename `auto_moderation.py` -> `automod.py`
